fix: prevent from crashing with duplicated columns
It was crashing when running it with 2 datasets that share a column name
(e.g. description). The fix is to use the record set name as part of the
column ID, similar to the examples we saw. It makes sense as the column
ID needs to be unique across the entire dataset.